Issue 915. Refactor getDateFormat to utils and remove unused dateFormats config.
