--- bases/rsptx/assignment_server_api/assignment_builder/src/config/dateFormats.ts
@@ -1,5 +1,0 @@
-export const getDateFormat = (): string => {
-  const lang = navigator.language;
-
-  return lang.startsWith("en") ? "mm/dd/yy" : "dd/mm/yy";
-};
